Compute the GPU Hessian eigendecomposition in real precision
analyzeHessianKernel read the (now double) Hessian blocks but did the symmetric
3x3 eigensolve, eigenvectors, curvature, anisotropy, and entropy in float, so a
double context lost precision in exactly the near-degenerate eigenvalue and
soft-mode regions that matter for normal modes. Convert eigenvalues_3x3_symmetric,
eigenvector_for_eigenvalue, the analyze kernel body, and the entropy reduction to
real (overloaded sqrt/cos/acos/fmin/fmax/log; M_PI literal cast to real; fmin/fmax
literals cast to avoid the double-mode overload ambiguity). The analysis output
buffers (eigenvalues, eigenvectors, curvature metrics, entropy, total entropy)
become mixed, allocated and read back at the context precision into double members;
the entropy-reduction shared memory and the getHessianAnalysis struct follow.
numNegative stays int.

single==mixed eigenvalues bit-identical; double agrees ~2.6e-6 and now carries the
double-precision eigendecomposition. test_precision_suite adds an eigendecomposition
precision case.

+++ platforms/cuda/src/CudaGridForceKernels.h
@@ -265,16 +265,16 @@ private:
     OpenMM::CudaArray numNegativeBuffer;         // [numAtoms] int
     OpenMM::CudaArray totalEntropyBuffer;        // [1] scalar
     bool analysisBuffersInitialized;             // Whether analysis buffers are allocated
-    std::vector<float> lastEigenvalues;
-    std::vector<float> lastEigenvectors;
-    std::vector<float> lastMeanCurvature;
-    std::vector<float> lastTotalCurvature;
-    std::vector<float> lastGaussianCurvature;
-    std::vector<float> lastFracAnisotropy;
-    std::vector<float> lastEntropy;
-    std::vector<float> lastMinEigenvalue;
+    std::vector<double> lastEigenvalues;
+    std::vector<double> lastEigenvectors;
+    std::vector<double> lastMeanCurvature;
+    std::vector<double> lastTotalCurvature;
+    std::vector<double> lastGaussianCurvature;
+    std::vector<double> lastFracAnisotropy;
+    std::vector<double> lastEntropy;
+    std::vector<double> lastMinEigenvalue;
     std::vector<int> lastNumNegative;
-    float lastTotalEntropy;
+    double lastTotalEntropy;
 
     bool skipGroupEnergyDownload_ = false;
 public:
